chore :: repository naming 통일
