+++ src/main/java/noteasy/sundo/queryfactory/persistmodel/classroom/factory/ClassRoomRepositoryCustom.java
@@ -4,6 +4,6 @@ import noteasy.sundo.queryfactory.persistmodel.classroom.ClassRoom;
 
 import java.util.Optional;
 
-public interface ClassRoomQueryFactoryNeed {
+public interface ClassRoomRepositoryCustom {
     Optional<ClassRoom> findByGradeAndClassNum(Integer grade, Integer classNum);
 }
