+++ src/main/java/noteasy/sundo/queryfactory/persistmodel/classroom/manager/ClassRoomPersistenceManager.java
@@ -2,8 +2,8 @@ package noteasy.sundo.queryfactory.persistmodel.classroom.manager;
 
 import noteasy.sundo.queryfactory.BaseQueryFactory;
 import noteasy.sundo.queryfactory.persistmodel.classroom.ClassRoom;
-import noteasy.sundo.queryfactory.persistmodel.classroom.factory.ClassRoomQueryFactoryNeed;
+import noteasy.sundo.queryfactory.persistmodel.classroom.factory.ClassRoomRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassRoomPersistenceManager extends JpaRepository<ClassRoom, Long>, BaseQueryFactory<ClassRoom, Long>, ClassRoomQueryFactoryNeed {
+public interface ClassRoomPersistenceManager extends JpaRepository<ClassRoom, Long>, BaseQueryFactory<ClassRoom, Long>, ClassRoomRepositoryCustom {
 }
